requirements.txt
requests>=2.31.0
pyyaml>=6.0
aiohttp>=3.9.0
urllib3>=2.0.0
python-dateutil>=2.8.2

